move MetaUnit props
